requirements.txt
pyside2         # For the QT5 GUI
code128         # See https://pypi.org/project/code128/
Pillow          # Python Image Library
pyinstaller     # To make the single executable
pywin32         # To make the .exe which pyinstaller makes actually work